backend/requirements.txt
django==4.2.27
djangorestframework==3.14.0
django-cors-headers==4.4.0
cloudinary==1.44.1
python-decouple==3.8
PyJWT==2.8.0
Pillow
psycopg2-binary
dj-database-url==2.1.0
whitenoise
gunicorn
